Fix t100 detection
the check for the decicion to install a t100 or t250 does not work. Issue is the use a the wrong return value. 

The code assume that the hostname check return an entry into the variable VALUE, but it gives it back into the variable debug.

Fixing the variable name from debug to VALUE fixes the detection.