feat: Add BTT Kraken V1.1 board
The only change from the 1.0 revision is a different sense resistor
installed for the primary four stepper drivers.

We intentionally don't default to the new version in the T250 template
so at worst a lower sense resistor is physically present (V1.0: 22 mOhm)
on the board than defined in our config (V1.1: 50 mOhm). The other way
around could lead to damaging currents getting supplied to the motors.